awesome updates suggested by manuelgit add .
